Baekjoon_1_8

+++ src/Baekjoon_1_5.java
@@ -8,8 +8,8 @@ public class Baekjoon_1_5
     // 문제 : 두 정수 A와 B를 입력받은 다음, A/B를 출력하는 프로그램을 작성하시오.
 
     Scanner sc = new Scanner(System.in);
-    int A = 0;
-    int B = 0;
+    double A = 0;
+    double B = 0;
     A = sc.nextInt();
     B = sc.nextInt();
     System.out.println(A/B);
